kexec_file: add mechanism to update kexec segments
kexec_update_segment allows a given segment in kexec_image to have its
contents updated.  This is useful if the current kernel wants to send
information to the next kernel that is up-to-date at the time of reboot.

Before modifying the segment the image checksum is verified, and after the
segment is updated the checksum is recalculated and updated in the kexec
image.

Link: http://lkml.kernel.org/r/1473900890-1476-5-git-send-email-bauerman@linux.vnet.ibm.com
Signed-off-by: Thiago Jung Bauermann <bauerman@linux.vnet.ibm.com>
Suggested-by: Mimi Zohar <zohar@linux.vnet.ibm.com>
Cc: Eric Biederman <ebiederm@xmission.com>
Cc: Dave Young <dyoung@redhat.com>
Cc: Vivek Goyal <vgoyal@redhat.com>
Cc: Baoquan He <bhe@redhat.com>
Cc: Michael Ellerman <mpe@ellerman.id.au>
Cc: Stewart Smith <stewart@linux.vnet.ibm.com>
Cc: Eric Richter <erichte@linux.vnet.ibm.com>
Cc: Balbir Singh <bsingharora@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>