ima: store the builtin/custom template definitions in a list
The builtin and single custom templates are currently stored in an array. 
In preparation for being able to restore a measurement list containing
multiple builtin/custom templates, this patch stores the builtin and
custom templates as a linked list.  This will permit defining more than
one custom template per boot.

Link: http://lkml.kernel.org/r/1474911029-6372-8-git-send-email-zohar@linux.vnet.ibm.com
Signed-off-by: Mimi Zohar <zohar@linux.vnet.ibm.com>
Acked-by: "Eric W. Biederman" <ebiederm@xmission.com>
Cc: Dave Young <dyoung@redhat.com>
Cc: Thiago Jung Bauermann <bauerman@linux.vnet.ibm.com>
Cc: Benjamin Herrenschmidt <benh@kernel.crashing.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>