mm: warn about allocations which stall for too long
Currently we do warn only about allocation failures but small allocations
are basically nofail and they might loop in the page allocator for a long
time.  Especially when the reclaim cannot make any progress - e.g. 
GFP_NOFS cannot invoke the oom killer and rely on a different context to
make a forward progress in case there is a lot memory used by filesystems.

Give us at least a clue when something like this happens and warn about
allocations which take more than 10s.  Print the basic allocation context
information along with the cumulative time spent in the allocation as well
as the allocation stack.  Repeat the warning after every 10 seconds so
that we know that the problem is permanent rather than ephemeral.

Link: http://lkml.kernel.org/r/20160929084407.7004-3-mhocko@kernel.org
Signed-off-by: Michal Hocko <mhocko@suse.com>
Cc: Vlastimil Babka <vbabka@suse.cz>
Cc: Tetsuo Handa <penguin-kernel@I-love.SAKURA.ne.jp>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Mel Gorman <mgorman@suse.de>
Cc: Dave Hansen <dave.hansen@intel.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>