ima: store the builtin/custom template definitions in a list
The builtin and single custom templates are currently stored in an array. 
In preparation for being able to restore a measurement list containing
multiple builtin/custom templates, this patch stores the builtin and
custom templates as a linked list.  This will permit defining more than
one custom template per boot.

Link: http://lkml.kernel.org/r/1477017898-10375-8-git-send-email-bauerman@linux.vnet.ibm.com
Signed-off-by: Mimi Zohar <zohar@linux.vnet.ibm.com>
Cc: "Eric W. Biederman" <ebiederm@xmission.com>
Cc: Andreas Steffen <andreas.steffen@strongswan.org>
Cc: Balbir Singh <bsingharora@gmail.com>
Cc: Dave Young <dyoung@redhat.com>
Cc: Josh Sklar <sklar@linux.vnet.ibm.com>
Cc: Michael Ellerman <mpe@ellerman.id.au>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>